requirements.txt
py-cord[voice]
pyyaml
python-dotenv